# ENH: Improve performance of tril_indices and triu_indices