--- a/benchmarks/benchmarks/bench_core.py
+++ b/benchmarks/benchmarks/bench_core.py
@@ -93,6 +93,12 @@ class Core(Benchmark):
     def time_tril_l10x10(self):
         np.tril(self.l10x10)
 
+    def time_triu_indices_500(self):
+        np.triu_indices(500)
+
+    def time_tril_indices_500(self):
+        np.tril_indices(500)
+
 
 class Temporaries(Benchmark):
     def setup(self):