`np.nonzero` seems slow compared to other methods. Use `np.indices` and `np.broadcast_to` to speed it up.

Closes #18153.

Performance comparison:
```python
import numpy as np


def tril_indices(n, k=0, m=None):
    tri = np.tri(n, m, k=k, dtype=bool)

    return tuple(np.broadcast_to(inds, tri.shape)[tri]
                 for inds in np.indices(tri.shape, sparse=True))


def triu_indices(n, k=0, m=None):
    tri = ~np.tri(n, m, k=k - 1, dtype=bool)

    return tuple(np.broadcast_to(inds, tri.shape)[tri]
                 for inds in np.indices(tri.shape, sparse=True))

# Original numpy implementation with nonzero:
%timeit np.tril_indices(500)
2.14 ms ± 52.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# New implementation:
%timeit tril_indices(500)
684 µs ± 60.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```