@Qiyu8 2021-01-18
you should compare new implements in `tril_indices` and `triu_indices` with `nonzero`,  Can you write a benchmark in `bench_core.py`?