@Qiyu8 2021-01-19
I don't think #18183 is related, can you add your crash logs here? From the CI perspective, The only failure is `TestMatmul.test_vector_matrix_values`.