@Qiyu8 2021-01-20
Can you provide the benchmark result by running `python runtests.py --bench-compare master bench_core.Core`?