@Qiyu8 2021-01-21
you can read the benchmark guide [here](https://github.com/numpy/numpy/tree/master/benchmarks).