@Illviljan 2021-01-21
@Qiyu8 It's still not working for me, got a little further though:
* `python runtests.py --bench` -Works
* `python runtests.py --bench-compare master bench_core` - does not work.

```
(base) C:\Users\J.W\Documents\GitHub\numpy>python runtests.py --bench-compare master bench_core
Traceback (most recent call last):
  File "C:\Users\J.W\Documents\GitHub\numpy\numpy\__init__.py", line 126, in <module>
    from numpy.__config__ import show as show_config
ModuleNotFoundError: No module named 'numpy.__config__'

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "runtests.py", line 642, in <module>
    main(argv=sys.argv[1:])
  File "runtests.py", line 172, in main
    _temp = __import__(PROJECT_MODULE)
  File "C:\Users\J.W\Documents\GitHub\numpy\numpy\__init__.py", line 131, in <module>
    raise ImportError(msg) from e
ImportError: Error importing numpy: you should not try to import numpy from
        its source directory; please exit the numpy source tree, and relaunch
        your python interpreter from there.
```

Also virtualenv doesn't seem to be the recommended way anymore:
> Note
> 
> If you are using Python 3.3 or newer, the venv module is the preferred way to create and manage virtual environments. venv is included in the Python standard library and requires no additional installation. If you are using venv, you may skip this section. 
> 
https://packaging.python.org/guides/installing-using-pip-and-virtual-environments/#installing-virtualenv